Split UnitOfWorkBuilder off of IncementalExecutionIntegrationTest
